requirements.txt
aniso8601==7.0.0
autopep8==1.5
Django==2.2
django-cors-headers==3.2.1
django-graphql-jwt==0.3.0
graphene==2.1.8
graphene-django==2.9.0
graphql-core==2.3.1
graphql-relay==2.0.1
gunicorn==20.0.4
promise==2.3
pycodestyle==2.5.0
PyJWT==1.7.1
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
